pull request student unclutter update and hosting setup
