+++ src/app/change-password/change-password.component.html
@@ -11,22 +11,20 @@
                 Please enter your email in the text area below so that
                 we will be able to help you in retrieving your account.
             </p>
-            <input  type="text" 
-                    class="form-control" 
-                    placeholder="Enter your email here."
-                    formControlName="inputEmail">
+            <input type="text" 
+                class="form-control" 
+                placeholder="Enter your email here."
+                formControlName="inputEmail">
         </div>
 
         <div class="container-footer">
-            <!-- <p *ngIf="sent">Message sent to "{{ email }}".</p>
-            <p *ngIf="failSent">Error. Cannot find "{{ email }}".</p> -->
             <button type="button" 
-                    class="btn btn-danger btn-sm"
-                    (click)="goToLogin()"> Cancel
+                class="btn btn-danger btn-sm"
+                (click)="goToLogin()"> Cancel
             </button>
             <button type="button" 
-                    class="btn btn-primary btn-sm"
-                    (click)="find()"> Find
+                class="btn btn-primary btn-sm"
+                (click)="find()"> Find
             </button>
         </div>
     </div>
