+++ src/app/log-in/log-in.component.html
@@ -5,7 +5,6 @@
 
     <!--Log In Form-->
     <form class="form-horizontal" [formGroup]="loginForm" (ngSubmit)="logIn()">
-
         <table>
             <tr>
                 <td>
@@ -14,7 +13,11 @@
                 <td>
                     <div class="form-group">
                         <div class="col-sm-10">
-                            <input type="text" class="form-control" formControlName="email" id="email" placeholder="Input your email here"
+                            <input type="text" 
+                                class="form-control" 
+                                formControlName="email" 
+                                id="email" 
+                                placeholder="Input your email here"
                                 required>
                         </div>
                     </div>
@@ -28,8 +31,12 @@
                 <td>
                     <div class="form-group">
                         <div class="col-sm-10">
-                            <input type="password" class="form-control" formControlName="password" id="password"
-                                placeholder="Password" required>
+                            <input type="password" 
+                                class="form-control" 
+                                formControlName="password" 
+                                id="password"
+                                placeholder="Password" 
+                                required>
                         </div>
                     </div>
                 </td>
@@ -54,7 +61,7 @@
                     <br>
                     <div class="linkers">
                         <p>Forgot password?</p>
-                        <p class="hyperlinks" (click)="userChangePassword()">Click here.</p>
+                        <p class="hyperlinks" (click)="recoverPassword()">Click here.</p>
                     </div>
                     <div class="linkers">
                         <p>No account yet?</p>
@@ -64,5 +71,4 @@
             </tr>
         </table>
     </form>
-    <!--Log In Form End-->
 </div>
