+++ src/app/student/specific/pages/specific-quest-map/index.ts
@@ -1,0 +1,3 @@
+export * from './specific-quest-map-navbar/specific-quest-map-navbar.component';
+export * from './subpages/map-chart/map-chart.component';
+export * from './subpages/summary/summary.component';
