+++ src/app/student/specific/pages/specific-quest-map/specific-quest-map-navbar/specific-quest-map-navbar.component.css
@@ -1,0 +1,30 @@
+.navbarLine {
+    content: ' ';
+    margin-top: -5px;
+    display: block;
+    width: 95%;
+    border: 2px solid #d0d0d0;
+    border-radius: 4px;
+    -webkit-border-radius: 4px;
+    -moz-border-radius: 4px;
+
+    box-shadow: inset 0 1px 1px rgba(0, 0, 0, .05);
+    -webkit-box-shadow: inset 0 1px 1px rgba(0, 0, 0, .05);
+    -moz-box-shadow: inset 0 1px 1px rgba(0, 0, 0, .05);
+}
+
+.qMapLink {
+    text-decoration: none;
+    color: black;
+    font-size: 30px;
+    margin: 0 auto 0 auto;
+}
+
+.navbar {
+    width: 95%;
+}
+
+.verticalBar {
+    border-left: 2px solid black;
+    height: 30px;
+}
