+++ src/app/student/specific/pages/specific-quest-map/specific-quest-map-navbar/specific-quest-map-navbar.component.html
@@ -1,0 +1,7 @@
+<div class="navbar">
+	<a class="qMapLink" [routerLink]="['./']">Quest</a>
+	<div class="verticalBar"></div>
+	<a class="qMapLink" [routerLink]="['./subpages/summary']">Summary</a>
+</div>
+
+<span class="navbarLine"></span>
