+++ src/app/student/specific/pages/specific-quest-map/specific-quest-map-navbar/specific-quest-map-navbar.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+	selector: 'app-specific-quest-map-navbar',
+	templateUrl: './specific-quest-map-navbar.component.html',
+	styleUrls: ['./specific-quest-map-navbar.component.css']
+})
+export class SpecificQuestMapNavbarComponent implements OnInit {
+
+	constructor() { }
+
+	ngOnInit() {
+	}
+
+}
