+++ src/app/student/specific/pages/specific-quest-map/specific-quest-map.component.css
@@ -1,61 +1,3 @@
-/*Header3 Design*/
-h3:after {
-    content:' ';
-    margin-top: 5px;
-    display:block;
-    width: 90%;
-    border:2px solid #d0d0d0;
-    border-radius:4px;
-    -webkit-border-radius:4px;
-    -moz-border-radius:4px;
-    box-shadow:inset 0 1px 1px rgba(0, 0, 0, .05);
-    -webkit-box-shadow:inset 0 1px 1px rgba(0, 0, 0, .05);
-    -moz-box-shadow:inset 0 1px 1px rgba(0, 0, 0, .05);
-}
-
-.modal-header h4 {
-    margin: 0 auto;
-}
-
-ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-}
-
-.square {
-    position: relative;
-    margin: 20px;
-    display: inline-block; /* could be float */
-    overflow: auto; /* UPDATE: if content may overflow square */
-}
-.sq-setter-w {
-    width: 100%;
-    height: auto;
-    visibility: hidden;
-}
-.sq-setter-h {
-    width: auto;
-    height: 100%;
-    visibility: hidden;
-}
-.sq-content {
-    position: absolute;
-    z-index: 1;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-}
-
-h4 {
-    text-align: center;
-}
-
-.quest-table tr td{
-    border: 1px solid #ddd;
-}
-
-.row-header{
-    font-weight: bold;
+.side-body {
+    padding-top: 25px;
 }
