+++ src/app/student/specific/pages/specific-quest-map/subpages/summary/summary.component.css
@@ -1,0 +1,3 @@
+.quest-table tr td{
+    border: 1px solid #ddd;
+}
