+++ src/app/student/specific/specific.module.ts
@@ -1,61 +1,73 @@
 //Core Imports 
 import {
-  NgModule
+	NgModule
 } from '@angular/core';
 
 //Application Imports
 import {
-  SharedModule
+	SharedModule
 } from 'shared/shared.module';
 
 import {
-  SpecificCharacterComponent,
-  SpecificMyCourseComponent,
-  SpecificNewsComponent,
-  SpecificProfileComponent,
-  SpecificQuestMapComponent
+	SpecificCharacterComponent,
+	SpecificMyCourseComponent,
+	SpecificNewsComponent,
+	SpecificProfileComponent,
+	SpecificQuestMapComponent
 } from 'student/specific/pages';
 
 import {
-  SpecificComponent
+	SpecificComponent
 } from 'student/specific/specific.component';
 
 import {
-  SpecificRoutingModule
+	SpecificRoutingModule
 } from 'student/specific/specific-routing.module';
 
 import {
-  SpecificSidetabComponent,
-  SpecificTopnavbarComponent
+	SpecificSidetabComponent,
+	SpecificTopnavbarComponent
 } from 'app/student/specific';
 
-
-import { 
-  GeneralRoutingModule 
+import {
+	GeneralRoutingModule
 } from 'student/general/general-routing.module';
 
-import { 
-  GeneralModule 
+import {
+	GeneralModule
 } from 'student/general/general.module';
 
+import { 
+	SpecificQuestMapNavbarComponent 
+} from './pages/specific-quest-map/specific-quest-map-navbar/specific-quest-map-navbar.component';
 
+import { 
+	MapChartComponent 
+} from './pages/specific-quest-map/subpages/map-chart/map-chart.component';
+
+import { 
+	SummaryComponent 
+} from './pages/specific-quest-map/subpages/summary/summary.component';
 
 @NgModule({
-  imports: [
-    SharedModule,
-    SpecificRoutingModule,
-    GeneralRoutingModule,
-    GeneralModule
-  ],
-  declarations: [
-    SpecificCharacterComponent,
-    SpecificComponent,
-    SpecificMyCourseComponent,
-    SpecificNewsComponent,
-    SpecificProfileComponent,
-    SpecificQuestMapComponent,
-    SpecificSidetabComponent,
-    SpecificTopnavbarComponent
-  ]
+	imports: [
+		SharedModule,
+		SpecificRoutingModule,
+		GeneralRoutingModule,
+		GeneralModule
+	],
+	declarations: [
+		SpecificCharacterComponent,
+		SpecificComponent,
+		SpecificMyCourseComponent,
+		SpecificNewsComponent,
+		SpecificProfileComponent,
+		SpecificQuestMapComponent,
+		SpecificSidetabComponent,
+		SpecificTopnavbarComponent,
+		SpecificQuestMapNavbarComponent,
+		MapChartComponent,
+		SummaryComponent
+	]
 })
 export class SpecificModule { }
