+++ src/app/teacher/specific/pages/specific-quest-map/subpages/map-chart/map-chart.component.css
@@ -2,74 +2,32 @@
     display: none;
 }
 
-ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-}
-
-/*Create Quest Modal*/
-.modal-header h4 {
-    margin: 0 auto;
-}
-
-.modal-body {
-    width: 100%;
-    padding-right: 0px;
-    padding-left: 0px;
-}
- 
-.modal-body .row {
-    padding-top: 5px;
-    padding-bottom: 5px;
-}
-
-.legendSymbol {
-    width: 8px;
-    height: 8px;
-    margin-left: 5px;
+/* Modal */
+.questDesc:after {
+    content: ' ';
+    margin: 10px auto 0 auto;
+    display: block;
+    width: 90%;
+    border: 2px solid #d0d0d0;
     border-radius: 4px;
-    display: inline-block;
 }
 
-.legend {
-    text-align: left;
-    float: left;
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: -5px;
+.modal-title {
+    margin: 0 auto 0 auto;
+    width: 100%;
+    text-align: center;
 }
 
-.questLegend {
-    margin-top: -30px;
-    margin-bottom: 15px;
-    width: 95%;
-    height: 25px;
+table {
+    width: 100%;
 }
 
-.newQuest {
-    border: 1px solid black;
-    background-color: black;
+td {
+    vertical-align: top;
 }
 
-.locked {
-    border: 1px solid gray;
-    background-color: gray;
-}
-
-.inProgress {
-    border: 1px solid green;
-    background-color: green;
-}
-
-.finished {
-    border: 1px solid orange;
-    background-color: orange;
-}
-
-.deadline {
-    border: 1px solid red;
-    background-color: red;
+ol {
+    list-style: disc;
 }
 
 @media screen and (max-width: 768px) {
