+++ src/app/teacher/specific/pages/specific-quest-map/subpages/summary/summary.component.css
@@ -3,6 +3,7 @@
     text-align: center;
 }
 
+/* Modal */
 .questDesc:after {
     content: ' ';
     margin: 10px auto 0 auto;
@@ -16,6 +17,18 @@
     margin: 0 auto 0 auto;
     width: 100%;
     text-align: center;
+}
+
+table {
+    width: 100%;
+}
+
+td {
+    vertical-align: top;
+}
+
+ol {
+    list-style: disc;
 }
 
 /* Summary table view */
@@ -59,17 +72,4 @@
 
 .btn:hover {
     cursor: pointer;
-}
-
-/* Modal table */
-table {
-    width: 100%;
-}
-
-td {
-    vertical-align: top;
-}
-
-ol {
-    list-style: disc;
 }
